New functions added
1) splitCells(inputfilename,cellsize,verbose)

Splits a LiDAR cloud file horizontally in tiles with cellsize (m) side.

2) slices(inputfname,nslices,verbose)

Clip a LiDAR cloud vertically in n slices.

3) toFloor(inputfname,outputfname,floor,verbose)

Genenate a new outputfilename LiDAR file with all z coords equal floor, that is, transform a LiDAR cloud in a 2 dimensions cloud.

4) topBottom(command,inputfname,outputfname,percent,verbose)

If command is 't', generates a new file containing a top slice of the inputfilename file. The slice will correspond to a 'percent' percent of the total height of the cloud.

If command is 'b', generates a file containing a bottom slice of the original cloud.